woundcare-ai/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0.post1
pydantic==2.5.3
python-dotenv==1.0.1
